Refactor layer helpers a little
