+++ src/lib/maplibre/index.ts
@@ -1,3 +1,3 @@
 export * from "./utils";
-export * from "./layer_helper"
-export * from "./road_layers"
+export * from "./layer_helper";
+export * from "./layer_helper_utils";
