+++ src/lib/maplibre/layer_helper.ts
@@ -1,21 +1,30 @@
 import type { Map } from "maplibre-gl";
 
 export class LayerHelper {
-    id: string;
-    defaultPaintValues: Array<[string, string]>;
+  id: string;
+  defaultPaintValues: Array<[string, string]>;
 
-    constructor(id: string, defaultPaintValues: Array<[string, string]>) {
-        this.id = id;
-        this.defaultPaintValues = defaultPaintValues
-    }
+  constructor(layer: any) {
+    this.id = layer.id;
+    this.defaultPaintValues = this.getLayerValues(layer);
+  }
 
-    setProperty(map: Map, propertyName: string, value: string) {
-        map.setPaintProperty(this.id, propertyName, value);
-    }
+  getLayerValues(layer: any): Array<[string, string]> {
+    return Object.keys(layer.paint ? layer.paint : {}).map(
+      (paintPropertyKey) => {
+        //@ts-ignore Not sure how to convince ts that layer.paint will contain strings
+        return [paintPropertyKey, layer.paint[paintPropertyKey]];
+      }
+    );
+  }
 
-    returnToDefaultPaintValues(map: Map) {
-        this.defaultPaintValues.forEach(([propertyName, value]) => {
-            map.setPaintProperty(this.id, propertyName, value);
-        });
-    }
+  setProperty(map: Map, propertyName: string, value: string) {
+    map.setPaintProperty(this.id, propertyName, value);
+  }
+
+  returnToDefaultPaintValues(map: Map) {
+    this.defaultPaintValues.forEach(([propertyName, value]) => {
+      map.setPaintProperty(this.id, propertyName, value);
+    });
+  }
 }
